#  Improve performance in where push the element into non shared Array …